Refactor PersonTable location and add tests
The PersonTable component has been moved to its own subdirectory. Alongside this change, tests for the component have been added, ensuring that proper data renders. Moreover, updates were made in personTestData.ts to provide mock data for testing purposes.
